+++ CHANGELOG.md
@@ -14,6 +14,7 @@ New functionality:
 * [OLMIS-6441](https://openlmis.atlassian.net/browse/OLMIS-6441): Added UI for managing program orderables.
 * [OLMIS-6476](https://openlmis.atlassian.net/browse/OLMIS-6476): Added sorting of orderables by product name.
 * [OLMIS-6488](https://openlmis.atlassian.net/browse/OLMIS-6488): Added orderables caching on login action.
+* [OLMIS-6419](https://openlmis.atlassian.net/browse/OLMIS-6419): Added FTAP caching on login action.
 
 Improvements:
 * [OLMIS-6389](https://openlmis.atlassian.net/browse/OLMIS-6389): Added caching of system notifications on Home page and in header indicator.